Dashboard/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-multipart==0.0.20
opencv-python==4.11.0.86
numpy==2.2.4
websockets==15.0.1
ultralytics==8.3.140
deep-sort-realtime==1.3.2
flake8==7.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
coverage==7.8.1
httpx==0.28.1
SQLAlchemy==2.0.41
psycopg2==2.9.10
python-dotenv==1.1.0
